Simplify Packaging (#44)
* Removes all `extras_require` from the `setup.py` and all import related hacks around the codebase.
* Remove SQLAlchemy as a possible install target, thus defaulting the app to Peewee.

## Motivation

I am removing all of this mess because of the effort to make this code base work without Peewee. In order to effectively do that, I would need to wrap all the Peewee related classes in an `if` statement. I don't like that at all.

But that did get me thinking about the motivation about this library. This library is supposed to be a repository of shared code for all Croscon projects so that core pieces of infrastructure, like the `ArrowDateTimeField`, such that when a change is made, it can be pushed out to all Croscon projects without a bunch of copying and pasting. At Croscon, we no longer build apps using SQLAlchemy, so why even support it? Let's default to Peewee for everything and not even think or qualify it.

This does come with one downside: We have to install both Arrow and Pendulum for this library to work. This will add just a little bloat, but I am comfortable with this change in the name of progress and ease of use.

Also, by removing all these extra requires, our Travis CI builds pass in a minute instead of 30.